Remove files apparently reponsible for displaying environment variables
We don't want the environment variables used by a build to be visible on that build's status page, because if they are our Perforce (and possibly other) passwords would be exposed to developers.  I don't really have any real idea how to stop that from happening, so this is a very crude first attempt to kill the display by removing the files which I think might possibly be responsible for it. It will be a miracle if this works.